Last commit before push to main - this is merged with main, tested, working version

+++ BlazorWASM-AppLogic/Application/DAOsInterfaces/IPostDao.cs
@@ -7,7 +7,7 @@ namespace Application.DAOsInterfaces;
 
 public interface IPostDao
 {
-	Task<int> CreateAsync(PostCreationDto postCreationDto);
+	Task<long> CreateAsync(PostCreationDto postCreationDto);
 	Task<IEnumerable<Post>> GetAsync(SearchPostParametersDto parametersDto);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
 }
