+++ BlazorWASM-AppLogic/Domain/DTOs/SearchPostTagParameters.cs
@@ -3,9 +3,9 @@ namespace Domain.DTOs;
 public class SearchPostTagParameters
 {
     public string? tagContains { get;}
-    public int? postId { get;}
+    public long? postId { get;}
 
-    public SearchPostTagParameters(string? tagContains = null, int? postId = null)
+    public SearchPostTagParameters(string? tagContains = null, long? postId = null)
     {
         this.tagContains = tagContains;
         this.postId = postId;
