+++ BlazorWASM-AppLogic/HttpClients/ClientInterfaces/ITagService.cs
@@ -5,7 +5,7 @@ namespace HttpClients.ClientInterfaces;
 
 public interface ITagService
 {
-    Task<ICollection<TagPost>> GetPostTagAsync(int? postId = null, string? tagContains = null);
+    Task<ICollection<TagPost>> GetPostTagAsync(long? postId = null, string? tagContains = null);
     Task<ICollection<Tag>> GetTagListAsync(string? tagContains = null);
     Task<string[]> CreateAsync(PostTagCreationDto postTagCreation);
 }
